Repair the null pointer exception caused by the need to judge the archive thawing

+++ storage/object.go
@@ -1453,17 +1453,19 @@ func (yig *YigStorage) DeleteObject(reqCtx RequestContext,
 		}
 	}
 
-	if object.StorageClass == ObjectStorageClassGlacier {
-		if bucket.Versioning == BucketVersioningDisabled ||
-			(bucket.Versioning == BucketVersioningEnabled && object.VersionId != "") {
-			freezer, err := yig.MetaStorage.GetFreezer(object.BucketName, object.Name, object.VersionId)
-			if err == nil {
-				err = yig.MetaStorage.DeleteFreezer(freezer)
-				if err != nil {
+	if object != nil {
+		if object.StorageClass == ObjectStorageClassGlacier {
+			if bucket.Versioning == BucketVersioningDisabled ||
+				(bucket.Versioning == BucketVersioningEnabled && object.VersionId != "") {
+				freezer, err := yig.MetaStorage.GetFreezer(object.BucketName, object.Name, object.VersionId)
+				if err == nil {
+					err = yig.MetaStorage.DeleteFreezer(freezer)
+					if err != nil {
+						return result, err
+					}
+				} else if err != ErrNoSuchKey {
 					return result, err
 				}
-			} else if err != ErrNoSuchKey {
-				return result, err
 			}
 		}
 	}
@@ -1652,17 +1654,19 @@ func (yig *YigStorage) DeleteObjects(reqCtx RequestContext, credential common.Cr
 
 	deleteFunc := func(object *meta.Object, tx driver.Tx) (result DeleteObjectResult, err error) {
 		bucketName, objectName, version := bucket.Name, object.Name, object.VersionId
-		if object.StorageClass == ObjectStorageClassGlacier {
-			if bucket.Versioning == BucketVersioningDisabled ||
-				(bucket.Versioning == BucketVersioningEnabled && object.VersionId != "") {
-				freezer, err := yig.MetaStorage.GetFreezer(object.BucketName, object.Name, object.VersionId)
-				if err == nil {
-					err = yig.MetaStorage.DeleteFreezer(freezer)
-					if err != nil {
+		if object != nil {
+			if object.StorageClass == ObjectStorageClassGlacier {
+				if bucket.Versioning == BucketVersioningDisabled ||
+					(bucket.Versioning == BucketVersioningEnabled && object.VersionId != "") {
+					freezer, err := yig.MetaStorage.GetFreezer(object.BucketName, object.Name, object.VersionId)
+					if err == nil {
+						err = yig.MetaStorage.DeleteFreezer(freezer)
+						if err != nil {
+							return result, err
+						}
+					} else if err != ErrNoSuchKey {
 						return result, err
 					}
-				} else if err != ErrNoSuchKey {
-					return result, err
 				}
 			}
 		}
